+++ R/schoenersD.R
@@ -4,15 +4,17 @@
 #' automatically normalize surfaces to sum to 1.
 #'
 #' Calculates similarity value of two RasterLayers using Schoener's D-metric.
-#' @param rast1 Input RasterLayer
-#' @param rast2 Input RasterLayer 2
+#' @param rast1 First object of class SpatRaster
+#' @param rast2 Second object of class SpatRaster
 #'
-#' @importFrom raster cellStats
+#'
+#' @importFrom terra rast
+#' @importFrom terra global
 #'
 #' @examples
 #' # Generate example probability surfaces.
-#' myiso <- raster::rasterFromXYZ(isoscape)
-#' myiso_sd <- raster::rasterFromXYZ(isoscape_sd)
+#' myiso <- rast(isoscape, type="xyz")
+#' myiso_sd <- rast(isoscape_sd, type="xyz")
 #' df <- data.frame(
 #'         ID = c(-100, -80, -50),
 #'         isotopeValue = c(-100, -80, -50),
@@ -39,14 +41,25 @@
 # Run pairwise comparisons.
 schoenersD <- function(rast1, rast2){
 
+  if(is(rast1, "raster")) rast1 <- terra::rast(rast1)
+  if(is(rast2, "raster")) rast2 <- terra::rast(rast2)
+
   stopifnot(
-    "Argument rast1 is not of class 'RasterLayer'" = class(rast1) == "RasterLayer" ,
-    "Argument rast2 is not of class 'RasterLayer'" = class(rast2) == "RasterLayer"
+    "Argument rast1 is not of class 'SpatRaster'" = is(rast1, "SpatRaster"),
+    "Argument rast2 is not of class 'SpatRaster'" = is(rast2, "SpatRaster")
   )
 
-  if( cellStats(rast1, sum) != 1 ) { rast1 <- .findProductThenNormalize(rast1) }
-  if( cellStats(rast2, sum) != 1 ) { rast2 <- .findProductThenNormalize(rast2) }
+  if( global(rast1, "sum") != 1 ) { rast1 <- .normprodrast(rast1) }
+  if( global(rast2, "sum") != 1 ) { rast2 <- .normprodrast(rast2) }
 
-  1 - (0.5 * cellStats(abs(rast1 - rast2), stat='sum'))
+  1 - (0.5 * global(abs(rast1 - rast2), "sum"))
 
   }
+
+
+
+
+
+
+
+